backend/pyproject.toml
[tool.poetry]
name = "ai-todo-chatbot-backend"
version = "0.1.0"
description = "AI-Powered Todo Chatbot Backend"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlmodel = "^0.0.16"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.6"
better-exceptions = "^0.3.3"
python-dotenv = "^1.0.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
openai = "^1.3.5"
google-generativeai = "^0.4.0"
httpx = "^0.25.2"
python-slugify = "^8.0.1"
aiofiles = "^23.2.1"
cryptography = "^41.0.8"
python-dateutil = "^2.8.2"
sqlalchemy = "^2.0.23"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/migrations/*"]